Insert vector of sql and bind variables per statement in table
This will be used to record the sql and corresponding bind variables.

When we reach or approach the limit of the statement table we can then
show the contents and use them to hopefully figure out where bugs happen
or where there are problems in our rule code.

The mixing of C++ into plain C code may seem somewhat strange, but there
are already precedents and needlessly falling back to error prone
pointer juggling to record this, especially while already debugging,
seems inappropriate as well.