Record sql and bind variables into the statement table
Adapt the logging function to show the contents.

Show the active number of statements in the table, and the current
limit. At the moment this limit is compile-time set to 50.

Show, somewhat redundantly, both the process-id and the process-id of the
parent process as this might facilitate debugging and cross referencing
the logging.