Further improve tracing and debugging SQL statement table
Use name sqlVec for the vector that records bind variables and the
corresponding sql.

Better standardize on "key:value" in logging statements to facilitate
simple grepping.

Add bind variables from the static array in low_level_odbc.cpp to the
tracing.

Indent the displaying of tracing statements in the logging to improve
direct human inspection.

Move displaying table after allocating statements to also be able to see
the first ones. That makes debugging (of the debugging!) easier for now.

As the overwhelming majority of statement table cases seems to be only
one level deep, we probably want to change that later.